pyk/tests: move simplifyBool tests from integration to unit tests
